+++ src/app/_ui/components/Header/Header.tsx
@@ -1,7 +1,8 @@
 import Link from "next/link";
 import style from "./Header.module.css";
 import { Each } from "../Each/Each";
-import { NavInfo, nav } from "@/app/_ui/components/Header/nav";
+import { NavInfo } from "@/app/types/type";
+import { nav } from "@/app/_ui/components/Header/nav";
 
 type Props = {
   open: Boolean;
