+++ src/app/types/type.ts
@@ -1,0 +1,11 @@
+export interface NavInfo {
+  name: string;
+  heading: string;
+  description: string;
+  sublinks: string[];
+};
+
+export interface BenefitInfo {
+  title: string;
+  text: string;
+};
